Redefine DBDiscoverySQLException's vendor codes
