+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-provider/shardingsphere-db-discovery-mysql/src/main/java/org/apache/shardingsphere/dbdiscovery/mysql/exception/mgr/InvalidMGRPluginException.java
@@ -15,18 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.dbdiscovery.mysql.exception;
+package org.apache.shardingsphere.dbdiscovery.mysql.exception.mgr;
 
+import org.apache.shardingsphere.dbdiscovery.mysql.exception.DBDiscoverySQLException;
 import org.apache.shardingsphere.infra.util.exception.external.sql.sqlstate.XOpenSQLState;
 
 /**
- * Invalid MGR replication group member exception.
+ * Invalid MGR plugin exception.
  */
-public final class InvalidMGRReplicationGroupMemberException extends DBDiscoverySQLException {
+public final class InvalidMGRPluginException extends DBDiscoverySQLException {
     
-    private static final long serialVersionUID = 3766206496827463146L;
+    private static final long serialVersionUID = 1703082683321097037L;
     
-    public InvalidMGRReplicationGroupMemberException(final String url, final String databaseName) {
-        super(XOpenSQLState.CHECK_OPTION_VIOLATION, 2, "`%s` is not in MGR replication group member in database `%s`", url, databaseName);
+    public InvalidMGRPluginException(final String databaseName) {
+        super(XOpenSQLState.GENERAL_ERROR, 80, "MGR plugin is not active in database `%s`", databaseName);
     }
 }
